requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
openai>=1.12.0
python-dotenv>=1.0.0
streamlit>=1.32.0
pyyaml>=6.0 